Make blue bar at the bottom of the black header full-width
Existing GOV.UK styles sites seem to have one of two styles in terms of the
blue bar

1. Exactly content width bar, just beneath the black header area (so there
   is white either side of the blue bar)

2. Full width blue bar beneath the black header area. This is common for
   documentation sites and sites that have tabs beneath the blue bar.

Before this change the blue bar was a 3rd style - a bit bigger than content
width, and also "in" the black header area, so with black either side of it.

This change makes the blue bar the full-width style, since this template is
more like 2. above, especially when tabs are enabled.

Note that to handle the cases of tabs enabled and tabs disabled, we have the
blue bar on both the tabs element, and the header. To avoid the visual
duplication, a negative margin is used. There doesn't seem to be a way of
avoiding this with the material design mkdocs HTML structure.